+++ tests/test_monitorcontrol.py
@@ -122,9 +122,7 @@ def test_luminance(
         monitor.set_luminance(original)
 
 
-@pytest.mark.skipif(
-    USE_ATTACHED_MONITORS, reason="not going to change your contrast"
-)
+@pytest.mark.skipif(USE_ATTACHED_MONITORS, reason="not going to change your contrast")
 def test_contrast(monitor: Monitor):
     contrast = monitor.get_contrast()
     contrast += 1
@@ -132,9 +130,7 @@ def test_contrast(monitor: Monitor):
     assert monitor.get_contrast() == contrast
 
 
-@pytest.mark.skipif(
-    USE_ATTACHED_MONITORS, reason="not going to turn off your monitors"
-)
+@pytest.mark.skipif(USE_ATTACHED_MONITORS, reason="not going to turn off your monitors")
 @pytest.mark.parametrize(
     "mode, expected",
     [
@@ -212,9 +208,7 @@ def test_input_source(
 @pytest.mark.skipif(USE_ATTACHED_MONITORS, reason="This is mocked")
 def test_get_input_source_type_c(monitor: Monitor):
     type_c_input = 27
-    with mock.patch.object(
-        monitor, "_get_vcp_feature", return_value=type_c_input
-    ):
+    with mock.patch.object(monitor, "_get_vcp_feature", return_value=type_c_input):
         try:
             monitor.get_input_source()
             assert 0, "Did not raise InputSourceValueError"
@@ -261,57 +255,64 @@ def test_convert_to_dict():
         "F7(00 01 02 03) F8(00 01) F9 EF FD(00 01) FE(00 01 02) FF"
     )
     expected = {
-        0x02: [],
-        0x04: [],
-        0x05: [],
-        0x08: [],
-        0x10: [],
-        0x12: [],
-        0x14: [0x05, 0x08, 0x0B],
-        0x16: [],
-        0x18: [],
-        0x1A: [],
-        0x52: [],
-        0x60: [0x11, 0x12, 0x0F, 0x10],
-        0xAC: [],
-        0xAE: [],
-        0xB2: [],
-        0xB6: [],
-        0xC0: [],
-        0xC6: [],
-        0xC8: [],
-        0xC9: [],
-        0xD6: [0x01, 0x04],
-        0xDF: [],
-        0x62: [],
-        0x8D: [],
-        0xF4: [],
-        0xF5: [0x00, 0x01, 0x02],
-        0xF6: [0x00, 0x01, 0x02],
-        0x4D: [],
-        0x4E: [],
-        0x4F: [],
-        0x15: [
-            0x01,
-            0x06,
-            0x09,
-            0x10,
-            0x11,
-            0x13,
-            0x14,
-            0x28,
-            0x29,
-            0x32,
-            0x44,
-            0x48,
-        ],
-        0xF7: [0x00, 0x01, 0x02, 0x03],
-        0xF8: [0x00, 0x01],
-        0xF9: [],
-        0xEF: [],
-        0xFD: [0x00, 0x01],
-        0xFE: [0x00, 0x01, 0x02],
-        0xFF: [],
+        0x02: {},
+        0x04: {},
+        0x05: {},
+        0x08: {},
+        0x10: {},
+        0x12: {},
+        0x14: {0x05: {}, 0x08: {}, 0x0B: {}},
+        0x16: {},
+        0x18: {},
+        0x1A: {},
+        0x52: {},
+        0x60: {0x11: {}, 0x12: {}, 0x0F: {}, 0x10: {}},
+        0xAC: {},
+        0xAE: {},
+        0xB2: {},
+        0xB6: {},
+        0xC0: {},
+        0xC6: {},
+        0xC8: {},
+        0xC9: {},
+        0xD6: {0x01: {}, 0x04: {}},
+        0xDF: {},
+        0x62: {},
+        0x8D: {},
+        0xF4: {},
+        0xF5: {0x00: {}, 0x01: {}, 0x02: {}},
+        0xF6: {0x00: {}, 0x01: {}, 0x02: {}},
+        0x4D: {},
+        0x4E: {},
+        0x4F: {},
+        0x15: {
+            0x01: {},
+            0x06: {},
+            0x09: {},
+            0x10: {},
+            0x11: {},
+            0x13: {},
+            0x14: {},
+            0x28: {},
+            0x29: {},
+            0x32: {},
+            0x44: {},
+            0x48: {},
+        },
+        0xF7: {0x00: {}, 0x01: {}, 0x02: {}, 0x03: {}},
+        0xF8: {0x00: {}, 0x01: {}},
+        0xF9: {},
+        0xEF: {},
+        0xFD: {0x00: {}, 0x01: {}},
+        0xFE: {0x00: {}, 0x01: {}, 0x02: {}},
+        0xFF: {},
     }
 
+    assert _convert_to_dict(caps_str) == expected
+
+
+def test_convert_to_dict_nested():
+    # https://github.com/newAM/monitorcontrol/issues/249
+    caps_str = "DC(00(00 12 13 14))"
+    expected = {0xDC: {0: {0: {}, 0x12: {}, 0x13: {}, 0x14: {}}}}
     assert _convert_to_dict(caps_str) == expected